"add new category" functionality

+++ src/main/java/polpapntua/multimediaproject2425/models/Priority.java
@@ -4,12 +4,13 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigInteger;
 
 public class Priority implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private BigInteger id;
     private String name;
     private int level;
     @JsonIgnore
@@ -22,8 +23,8 @@ public class Priority implements Serializable {
         this.level = level;
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public BigInteger getId() { return id; }
+    public void setId(BigInteger id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
