+++ src/main/java/polpapntua/multimediaproject2425/models/Reminder.java
@@ -1,9 +1,10 @@
 package polpapntua.multimediaproject2425.models;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 
 public class Reminder {
-    public Long id;
+    public BigInteger id;
     public Task task;
     public LocalDate reminderDate;
